Внесены изменения во 2 задачу
