+++ even_numb.txt
@@ -1,0 +1,1 @@
+2, 4, 56, 54, 54, 32, 88, 98, 234, 76,
